+++ met4j-graph/src/test/java/fr/inrae/toulouse/metexplore/met4j_graph/TestExportGraph.java
@@ -537,7 +537,8 @@ public class TestExportGraph {
         StringWriter w = new StringWriter();
         AttributeExporter attExport = AttributeExporter.full(bn)
                 .exportEdgeAttribute("Test", e -> e.getV1().getId())
-                .exportNodeAttribute("Test2", BioEntity::getName);
+                .exportNodeAttribute("Test2", BioEntity::getName)
+                .exportNodeAttribute("Test3", v->42.0);
         ExportGraph<BioEntity, BipartiteEdge, BipartiteGraph> export = new ExportGraph<>(bg, attExport);
         export.toGml(w);
         String output = w.toString();
@@ -549,6 +550,7 @@ public class TestExportGraph {
         Pattern NameAttRegex = Pattern.compile("Name\\s+\"\\w\"");
         Pattern ReversibleAttRegex = Pattern.compile("Reversible\\s+\"(true|false)\"");
         Pattern Test2AttRegex = Pattern.compile("Test2\\s+\"\\w\"");
+        Pattern Test3AttRegex = Pattern.compile("Test3\\s+42.0");
         Pattern TestAttRegex = Pattern.compile("Test\\s+\"[a-z][0-9]?\"");
         Pattern TypeAttRegex = Pattern.compile("Type\\s+\"(Compound|Reaction)\"");
         Pattern CompAttRegex = Pattern.compile("Compartment\\s+\"comp[2,3]?(,comp[2,3])?\"");
@@ -562,6 +564,7 @@ public class TestExportGraph {
         assertEquals("anomaly in Bipartite name attributes",14, NameAttRegex.matcher(output).results().count());
         assertEquals("anomaly in Bipartite reversibility attributes",7, ReversibleAttRegex.matcher(output).results().count());
         assertEquals("anomaly in Bipartite test attributes",14, Test2AttRegex.matcher(output).results().count());
+        assertEquals("anomaly in Bipartite test attributes",14, Test3AttRegex.matcher(output).results().count());
         assertEquals("anomaly in Bipartite test attributes",22, TestAttRegex.matcher(output).results().count());
         assertEquals("anomaly in Bipartite type attributes",14, TypeAttRegex.matcher(output).results().count());
         assertEquals("anomaly in Bipartite transport attributes",7, TranspAttRegex.matcher(output).results().count());
@@ -573,7 +576,11 @@ public class TestExportGraph {
         StringWriter w = new StringWriter();
         AttributeExporter attExport = AttributeExporter.full(bn)
                 .exportEdgeAttribute("Test", e -> e.getV1().getId())
-                .exportNodeAttribute("Test2", BioEntity::getName);
+                .exportNodeAttribute("Test2", BioEntity::getName)
+                .exportNodeAttribute("Test3", e -> "[\"toto\",\"tata\"]")
+                .exportNodeAttribute("Type", (v->{if (v instanceof BioMetabolite) return "Composé"; //override default att
+                    else if (v instanceof BioReaction) return "Reaction";
+                    else return v.getClass().getSimpleName();}));
         ExportGraph<BioEntity, BipartiteEdge, BipartiteGraph> export = new ExportGraph<>(bg, attExport);
         try {
             export.toJSONgraph(w);
@@ -586,8 +593,9 @@ public class TestExportGraph {
             Pattern NameAttRegex = Pattern.compile("\"Name\":\"\\w\"");
             Pattern ReversibleAttRegex = Pattern.compile("\"Reversible\":\"(true|false)\"");
             Pattern Test2AttRegex = Pattern.compile("\"Test2\":\"\\w\"");
+            Pattern Test3AttRegex = Pattern.compile("\"Test3\":\\[\"toto\",\"tata\"]");
             Pattern TestAttRegex = Pattern.compile("\"Test\":\"[a-z][0-9]?\"");
-            Pattern TypeAttRegex = Pattern.compile("\"Type\":\"(Compound|Reaction)\"");
+            Pattern TypeAttRegex = Pattern.compile("\"Type\":\"(Composé|Reaction)\"");
             Pattern CompAttRegex = Pattern.compile("\"Compartment\":\"comp[2,3]?(,comp[2,3])?\"");
             Pattern TranspAttRegex = Pattern.compile("Transport\":\"(true||false)\"");
 
@@ -595,6 +603,7 @@ public class TestExportGraph {
             assertEquals("anomaly in Bipartite EC attributes",7, ECAttRegex.matcher(output).results().count());
             assertEquals("anomaly in Bipartite Formula attributes",7, FormulaAttRegex.matcher(output).results().count());
             assertEquals("anomaly in Bipartite mass attributes",7, MassAttRegex.matcher(output).results().count());
+            assertEquals("anomaly in Bipartite list attributes",14, Test3AttRegex.matcher(output).results().count());
             assertEquals("anomaly in Bipartite compartment attributes",7, CompAttRegex.matcher(output).results().count());
             assertEquals("anomaly in Bipartite name attributes",14, NameAttRegex.matcher(output).results().count());
             assertEquals("anomaly in Bipartite reversibility attributes",7, ReversibleAttRegex.matcher(output).results().count());