+++ met4j-toolbox/src/test/java/integration/Tab2SbmlIT.java
@@ -2,16 +2,13 @@ package integration;
 
 import org.junit.Test;
 import utils.IThelper;
-import org.junit.Assert;
 
-import java.util.regex.Pattern;
-import java.util.regex.Matcher;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.io.IOException;
 import java.io.BufferedReader;
 import java.io.FileReader;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.regex.Pattern;
 
 import static org.junit.Assert.*;
 
@@ -28,7 +25,7 @@ public class Tab2SbmlIT {
             tmpDir.toFile().deleteOnExit();
         } catch (IOException e1) {
             e1.printStackTrace();
-            Assert.fail("Creation of the temporary directory");
+            fail("Creation of the temporary directory");
         }
         
         String inFile =IThelper.copyProjectResource("toy_model.tsv",tmpDir);
@@ -74,7 +71,7 @@ public class Tab2SbmlIT {
             tmpDir.toFile().deleteOnExit();
         } catch (IOException e1) {
             e1.printStackTrace();
-            Assert.fail("Creation of the temporary directory");
+            fail("Creation of the temporary directory");
         }
         
         String inFile =IThelper.copyProjectResource("toy_model_otherColumns.tsv",tmpDir);
@@ -124,7 +121,7 @@ public class Tab2SbmlIT {
             tmpDir.toFile().deleteOnExit();
         } catch (IOException e1) {
             e1.printStackTrace();
-            Assert.fail("Creation of the temporary directory");
+            fail("Creation of the temporary directory");
         }
         
         String inFile =IThelper.copyProjectResource("toy_model_otherSigns.tsv",tmpDir);
@@ -174,19 +171,17 @@ public class Tab2SbmlIT {
             tmpDir.toFile().deleteOnExit();
         } catch (IOException e1) {
             e1.printStackTrace();
-            Assert.fail("Creation of the temporary directory");
+            fail("Creation of the temporary directory");
         }
         
-        String inFile =IThelper.copyProjectResource("toy_model.tsv",tmpDir);
+        String inFile =IThelper.copyProjectResource("toy_model_cobra.tsv",tmpDir);
         IThelper.ProcessResult result = IThelper.runCli(
             "fr.inrae.toulouse.metexplore.met4j_toolbox.convert.Tab2Sbml",
             "-i",
             inFile,
-            "-rp",
-            "-mp",
-            "-e",
-            "_e",
-            "-cpt",
+            "-M_c",
+            "-b",
+            "e",
             "-o",
             actualOutput.toString()
         );
@@ -194,24 +189,36 @@ public class Tab2SbmlIT {
         
         BufferedReader reader = new BufferedReader(new FileReader(actualOutput.toFile()));
         String line = reader.readLine();
-        Pattern pattern0 = Pattern.compile(".*boundaryCondition=.true.*");
-        Pattern pattern1 = Pattern.compile(".*id=.M_A_ext_e.*");
+        Pattern pattern0 = Pattern.compile(".*<compartment[^>]+id=\"e\".*");
+        Pattern pattern1 = Pattern.compile(".*id=.M_A_e.*");
         Pattern pattern2 = Pattern.compile(".*id=.R_reac2.*");
+        Pattern pattern3 = Pattern.compile(".*reaction .*");
+        Pattern pattern4 = Pattern.compile(".*species .*");
+        Pattern pattern5 = Pattern.compile(".*boundaryCondition\s*=\s*\"true\".*");
         int nbMatch0 = 0;
         int nbMatch1 = 0;
         int nbMatch2 = 0;
+        int nbMatch3 = 0;
+        int nbMatch4 = 0;
+        int nbMatch5 = 0;
         while (line != null) {
             if(pattern0.matcher(line).matches()) nbMatch0++;
             if(pattern1.matcher(line).matches()) nbMatch1++;
             if(pattern2.matcher(line).matches()) nbMatch2++;
+            if(pattern3.matcher(line).matches()) nbMatch3++;
+            if(pattern4.matcher(line).matches()) nbMatch4++;
+            if(pattern5.matcher(line).matches()) nbMatch5++;
             line = reader.readLine();
         }
         reader.close();
-        assertEquals(3,nbMatch0);
+        assertEquals(1,nbMatch0);
         assertEquals(1,nbMatch1);
         assertEquals(1,nbMatch2);
+        assertEquals(7,nbMatch3);
+        assertEquals(8,nbMatch4);
+        assertEquals(3,nbMatch5);
         assertTrue(IThelper.isValidXml(actualOutput.toFile()));
-    
+
 
         assertEquals(0, result.exitCode());
         assertTrue(Files.exists(actualOutput));
@@ -229,7 +236,7 @@ public class Tab2SbmlIT {
             tmpDir.toFile().deleteOnExit();
         } catch (IOException e1) {
             e1.printStackTrace();
-            Assert.fail("Creation of the temporary directory");
+            fail("Creation of the temporary directory");
         }
         
         String inFile =IThelper.copyProjectResource("toy_model.tsv",tmpDir);
@@ -260,5 +267,90 @@ public class Tab2SbmlIT {
         assertEquals(0, result.exitCode());
         assertTrue(Files.exists(actualOutput));
     }
+
+    @Test
+    public void testTab2Sbml5() throws Exception {
+
+        Path actualOutput = null;
+        Path tmpDir = null;
+        try {
+            actualOutput = Files.createTempFile("test-output-Tab2Sbml", ".sbml");
+            tmpDir = Files.createTempDirectory("test-input-Tab2Sbml");
+            tmpDir.toFile().deleteOnExit();
+        } catch (IOException e1) {
+            e1.printStackTrace();
+            fail("Creation of the temporary directory");
+        }
+
+        String inFile =IThelper.copyProjectResource("toy_model_wrong.tsv",tmpDir);
+        IThelper.ProcessResult result = IThelper.runCli(
+                "fr.inrae.toulouse.metexplore.met4j_toolbox.convert.Tab2Sbml",
+                "-i",
+                inFile,
+                "-id",
+                "myModel",
+                "--ignore-failed-read",
+                "-o",
+                actualOutput.toString()
+        );
+
+
+        BufferedReader reader = new BufferedReader(new FileReader(actualOutput.toFile()));
+        String line = reader.readLine();
+        Pattern pattern0 = Pattern.compile(".*reaction .*");
+        Pattern pattern1 = Pattern.compile(".*species .*");
+        Pattern pattern2 = Pattern.compile(".*compartment .*");
+        Pattern pattern3 = Pattern.compile(".*id=.A_x.*");
+        int nbMatch0 = 0;
+        int nbMatch1 = 0;
+        int nbMatch2 = 0;
+        int nbMatch3 = 0;
+        while (line != null) {
+            System.out.println(line);
+            if(pattern0.matcher(line).matches()) nbMatch0++;
+            if(pattern1.matcher(line).matches()) nbMatch1++;
+            if(pattern2.matcher(line).matches()) nbMatch2++;
+            if(pattern3.matcher(line).matches()) nbMatch3++;
+            line = reader.readLine();
+        }
+        reader.close();
+        assertEquals(3,nbMatch2);
+        assertEquals(10,nbMatch0);
+        assertEquals(9,nbMatch1);
+        assertEquals(1,nbMatch3);
+        assertTrue(IThelper.isValidXml(actualOutput.toFile()));
+
+
+        assertEquals(0, result.exitCode());
+        assertTrue(Files.exists(actualOutput));
+    }
+
+    @Test
+    public void testTab2Sbml6() throws Exception {
+
+        Path actualOutput = null;
+        Path tmpDir = null;
+        try {
+            actualOutput = Files.createTempFile("test-output-Tab2Sbml", ".sbml");
+            tmpDir = Files.createTempDirectory("test-input-Tab2Sbml");
+            tmpDir.toFile().deleteOnExit();
+        } catch (IOException e1) {
+            e1.printStackTrace();
+            fail("Creation of the temporary directory");
+        }
+
+        String inFile =IThelper.copyProjectResource("toy_model_wrong.tsv",tmpDir);
+        IThelper.ProcessResult result = IThelper.runCli(
+                "fr.inrae.toulouse.metexplore.met4j_toolbox.convert.Tab2Sbml",
+                "-i",
+                inFile,
+                "-id",
+                "myModel",
+                "-o",
+                actualOutput.toString()
+        );
+        assertEquals(1, result.exitCode());
+    }
+
     
 }