+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.2.1
+[met4j-toolbox] New graph export format compatible with MetExploreViz
+
+[met4j-io] fix duplicated compartment suffix in Tab2Sbml and Tab2BioNetwork
+[met4j-toolbox] remove deprecated options in Tab2Sbml, avoid reliance on identifiers convention such as compartment suffix
+
 ## 2.2.0
 
 [met4j-toolbox] Fix Galaxy compatibility issues with SideCompoundScan, improve logging and criteria for side compound detection